+++ e2e-test/cypress/integration/project_permissions_spec.js
@@ -2,11 +2,13 @@ import _cloneDeep from "lodash/cloneDeep";
 
 const executingUser = { id: "mstein", displayname: "Mauro Stein" };
 const testUser = { id: "thouse", displayname: "Tom House" };
-let projectId;
-let permissionsBeforeTesting;
+let projectId, permissionsBeforeTesting, baseUrl, apiRoute;
 
 describe("Project Permissions", function() {
   before(() => {
+    baseUrl = Cypress.env("API_BASE_URL") || `${Cypress.config("baseUrl")}/test`;
+    apiRoute = baseUrl.toLowerCase().includes("test") ? "/test/api" : "/api";
+
     cy.login();
     cy.createProject("p-subp-assign", "subproject assign test").then(({ id }) => {
       projectId = id;
@@ -28,7 +30,7 @@ describe("Project Permissions", function() {
     intentsToRevoke.forEach(intent => cy.revokeProjectPermission(projectId, intent, userId));
   }
 
-  function checkPermissionsEquality(permissionsBeforeTesting, projectId) {
+  function assertUnchangedPermissions(permissionsBeforeTesting, projectId) {
     cy.listProjectPermissions(projectId).then(permissions => {
       expect(permissions).to.deep.equal(permissionsBeforeTesting.project);
     });
@@ -103,7 +105,7 @@ describe("Project Permissions", function() {
       .find("button[data-test^='pp-button']")
       .click();
     cy.get("[data-test=permission-close]").click();
-    checkPermissionsEquality(permissionsBeforeTesting, projectId);
+    assertUnchangedPermissions(permissionsBeforeTesting, projectId);
   });
 
   it("Submitting the permission dialog without any changes doesn't revoke nor grant permissions and close the dialog", function() {
@@ -111,8 +113,9 @@ describe("Project Permissions", function() {
       .find("button[data-test^='pp-button']")
       .click();
     cy.get("[data-test=permission-submit]").click();
+    // Confirmation opens
     cy.get("[data-test=permission-container]").should("not.be.visible");
-    checkPermissionsEquality(permissionsBeforeTesting, projectId);
+    assertUnchangedPermissions(permissionsBeforeTesting, projectId);
   });
 
   it("Submitting the permission dialog after adding a user opens a confirmation dialog", function() {
@@ -125,6 +128,7 @@ describe("Project Permissions", function() {
       .click()
       .type("{esc}");
     cy.get("[data-test=permission-submit]").click();
+    // Confirmation opens
     cy.get("[data-test=confirmation-dialog-cancel]").should("be.visible");
   });
 
@@ -139,6 +143,7 @@ describe("Project Permissions", function() {
       .click()
       .type("{esc}");
     cy.get("[data-test=permission-submit]").click();
+    // Confirmation opens
     cy.get("[data-test=confirmation-dialog-cancel]").should("be.visible");
   });
 
@@ -154,6 +159,7 @@ describe("Project Permissions", function() {
     // Remove permission
     changePermissionInGui("project.intent.grantPermission", testUser.id);
     cy.get("[data-test=permission-submit]").click();
+    // Confirmation opens
     cy.get("[data-test=confirmation-dialog-cancel]").should("not.be.visible");
 
     // Reset permissions
@@ -174,6 +180,7 @@ describe("Project Permissions", function() {
     // Add permission
     changePermissionInGui("project.intent.grantPermission", testUser.id);
     cy.get("[data-test=permission-submit]").click();
+    // Confirmation opens
     cy.get("[data-test=confirmation-dialog-cancel]").should("not.be.visible");
 
     // Reset permissions
@@ -206,6 +213,7 @@ describe("Project Permissions", function() {
     // Add permission
     changePermissionInGui("project.update", testUser.id);
     cy.get("[data-test=permission-submit]").click();
+    // Confirmation opens
     cy.get("[data-test=actions-table-body]")
       .should("be.visible")
       .children()
@@ -219,6 +227,7 @@ describe("Project Permissions", function() {
     // Add permission
     changePermissionInGui("project.intent.revokePermission", testUser.id);
     cy.get("[data-test=permission-submit]").click();
+    // Confirmation opens
     actionTableIncludes("view permissions");
   });
 
@@ -229,6 +238,7 @@ describe("Project Permissions", function() {
     // Add permission
     changePermissionInGui("project.viewDetails", testUser.id);
     cy.get("[data-test=permission-submit]").click();
+    // Confirmation opens
     cy.get("[data-test=actions-table-body]")
       .should("be.visible")
       .children()
@@ -245,16 +255,20 @@ describe("Project Permissions", function() {
     // Add permission
     changePermissionInGui("project.intent.revokePermission", testUser.id);
     cy.get("[data-test=permission-submit]").click();
+    // Confirmation opens
+    // listPermissions calls are done
     cy.get("[data-test=actions-table-body]")
       .should("be.visible")
       .children()
       .should("have.length", 3);
-    cy.get("[data-test=confirmation-dialog-confirm]")
-      .click()
-      .should("not.be.disabled", { timeout: 30000 })
-      .then(() => {
-        checkPermissionsEquality(addViewPermissions(permissionsBeforeTesting, testUser.id), projectId);
-      });
+    // Make sure cypress waits for future listPermissions calls
+    cy.server();
+    cy.route("GET", apiRoute + "/project.intent.listPermissions*").as("listPermissions");
+    cy.get("[data-test=confirmation-dialog-confirm]").click();
+    // Additional actions are executed
+    cy.wait("@listPermissions");
+    cy.get("[data-test=confirmation-dialog-confirm]").should("not.be.disabled");
+    assertUnchangedPermissions(addViewPermissions(permissionsBeforeTesting, testUser.id), projectId);
 
     // Reset permissions
     Cypress.Promise.all([
@@ -270,18 +284,21 @@ describe("Project Permissions", function() {
       .click();
     changePermissionInGui("project.assign", testUser.id);
     cy.get("[data-test=permission-submit]").click();
+    // Confirmation opens
     actionTableIncludes("view permissions");
     cy.get("[data-test=confirmation-dialog-cancel]").click();
     changePermissionInGui("project.assign", testUser.id);
     // Check grant
     changePermissionInGui("project.intent.grantPermission", testUser.id);
     cy.get("[data-test=permission-submit]").click();
+    // Confirmation opens
     actionTableIncludes("view permissions");
     cy.get("[data-test=confirmation-dialog-cancel]").click();
     changePermissionInGui("project.intent.grantPermission", testUser.id);
     // Check revoke
     changePermissionInGui("project.intent.revokePermission", testUser.id);
     cy.get("[data-test=permission-submit]").click();
+    // Confirmation opens
     actionTableIncludes("view permissions");
   });
 
@@ -294,32 +311,32 @@ describe("Project Permissions", function() {
     changePermissionInGui("project.viewDetails", testUser.id);
     changePermissionInGui("project.intent.listPermissions", testUser.id);
     cy.get("[data-test=permission-submit]").click();
+    // Confirmation opens
     cy.get("[data-test=actions-table-body]").should("be.visible");
-    cy.get("[data-test=confirmation-dialog-confirm]")
-      .click()
-      .should("be.not.disabled", { timeout: 30000 })
-      .click();
+    cy.server();
+    cy.route("GET", apiRoute + "/project.intent.listPermissions*").as("listPermissions");
+    cy.get("[data-test=confirmation-dialog-confirm]").click();
+    // Additional actions are executed
+    cy.wait("@listPermissions");
+    cy.get("[data-test=confirmation-dialog-confirm]").click();
+    // Original actions are executed
     cy.get("[data-test=permission-submit]").should("not.be.visible");
-    cy.get("[data-test=loading-indicator]")
-      // Wait until all permissions are granted
-      .should("not.be.visible", { timeout: 30000 })
-      .then(() => {
-        let permissions = addViewPermissions(permissionsBeforeTesting, testUser.id);
-        permissions.project["project.update"].push(testUser.id);
-        checkPermissionsEquality(permissions, projectId);
+    cy.wait("@listPermissions");
+    let permissions = addViewPermissions(permissionsBeforeTesting, testUser.id);
+    permissions.project["project.update"].push(testUser.id);
+    assertUnchangedPermissions(permissions, projectId);
 
-        // Reset permissions
-        Cypress.Promise.all([
-          cy.revokeProjectPermission(projectId, "project.viewSummary", testUser.id),
-          cy.revokeProjectPermission(projectId, "project.viewDetails", testUser.id),
-          cy.revokeProjectPermission(projectId, "project.update", testUser.id),
-          cy.revokeProjectPermission(projectId, "project.intent.listPermissions", testUser.id)
-        ]);
-      });
+    // Reset permissions
+    Cypress.Promise.all([
+      cy.revokeProjectPermission(projectId, "project.viewSummary", testUser.id),
+      cy.revokeProjectPermission(projectId, "project.viewDetails", testUser.id),
+      cy.revokeProjectPermission(projectId, "project.update", testUser.id),
+      cy.revokeProjectPermission(projectId, "project.intent.listPermissions", testUser.id)
+    ]);
   });
 
   it("Confirmation of multiple revoke permission changes grants permissions correctly", function() {
-    let permissionsCopy;
+    let permCopy;
 
     Cypress.Promise.all([
       cy.grantProjectPermission(projectId, "project.viewSummary", testUser.id),
@@ -329,7 +346,7 @@ describe("Project Permissions", function() {
     ]).then(() => {
       cy.listProjectPermissions(projectId).then(permissions => {
         permissionsBeforeTesting.project = permissions;
-        permissionsCopy = _cloneDeep(permissionsBeforeTesting);
+        permCopy = _cloneDeep(permissionsBeforeTesting);
 
         cy.get(`[data-test=project-card-${projectId}]`)
           .find("button[data-test^='pp-button']")
@@ -340,24 +357,20 @@ describe("Project Permissions", function() {
         changePermissionInGui("project.viewDetails", testUser.id);
         changePermissionInGui("project.intent.listPermissions", testUser.id);
         cy.get("[data-test=permission-submit]").click();
+        // Confirmation opens
+        cy.get("[data-test=confirmation-dialog-confirm]").should("be.visible");
+        cy.server();
+        cy.route("GET", apiRoute + "/project.intent.listPermissions*").as("listPermissions");
         cy.get("[data-test=confirmation-dialog-confirm]").click();
+        // Original actions are executed
         cy.get("[data-test=permission-submit]").should("not.be.visible");
-        cy.get("[data-test=loading-indicator]")
-          // Wait until all permissions are granted
-          .should("not.be.visible", { timeout: 30000 })
-          .then(() => {
-            // Equal permissions
-            permissionsCopy.project = removePermission(permissionsCopy.project, "project.update", testUser.id);
-            permissionsCopy.project = removePermission(permissionsCopy.project, "project.viewSummary", testUser.id);
-            permissionsCopy.project = removePermission(permissionsCopy.project, "project.viewDetails", testUser.id);
-            permissionsCopy.project = removePermission(
-              permissionsCopy.project,
-              "project.intent.listPermissions",
-              testUser.id
-            );
-
-            checkPermissionsEquality(permissionsCopy, projectId);
-          });
+        cy.wait("@listPermissions");
+        // Equal permissions
+        permCopy.project = removePermission(permCopy.project, "project.update", testUser.id);
+        permCopy.project = removePermission(permCopy.project, "project.viewSummary", testUser.id);
+        permCopy.project = removePermission(permCopy.project, "project.viewDetails", testUser.id);
+        permCopy.project = removePermission(permCopy.project, "project.intent.listPermissions", testUser.id);
+        assertUnchangedPermissions(permCopy, projectId);
       });
     });
   });