+++ frontend/src/reducers.js
@@ -3,24 +3,22 @@
  * If we were to do this in store.js, reducers wouldn't be hot reloadable.
  */
 
+import { connectRouter, LOCATION_CHANGE } from "connected-react-router";
 import { fromJS } from "immutable";
 import { combineReducers } from "redux-immutable";
-import { LOCATION_CHANGE, connectRouter } from "connected-react-router";
-
-import navbarReducer from "./pages/Navbar/reducer";
-import overviewReducer from "./pages/Overview/reducer";
-import subProjectReducer from "./pages/SubProjects/reducer";
-import dashboardReducer from "./pages/Dashboard/reducer";
-import notificationsReducer from "./pages/Notifications/reducer";
-import workflowReducer from "./pages/Workflows/reducer";
-import workflowitemDetailsReducer from "./pages/Workflows/WorkflowitemHistoryTab/reducer";
-import loginReducer from "./pages/Login/reducer";
-import documentsReducer from "./pages/Documents/reducer";
-import loadingReducer from "./pages/Loading/reducer";
-import userDashboardReducer from "./pages/Users/reducer";
-import nodeDashboardReducer from "./pages/Nodes/reducer";
 import analyticsReducer from "./pages/Analytics/reducer";
 import confirmationReducer from "./pages/Confirmation/reducer";
+import documentsReducer from "./pages/Documents/reducer";
+import loadingReducer from "./pages/Loading/reducer";
+import loginReducer from "./pages/Login/reducer";
+import navbarReducer from "./pages/Navbar/reducer";
+import nodeDashboardReducer from "./pages/Nodes/reducer";
+import notificationsReducer from "./pages/Notifications/reducer";
+import overviewReducer from "./pages/Overview/reducer";
+import subProjectReducer from "./pages/SubProjects/reducer";
+import userDashboardReducer from "./pages/Users/reducer";
+import workflowReducer from "./pages/Workflows/reducer";
+import workflowitemDetailsReducer from "./pages/Workflows/WorkflowitemHistoryTab/reducer";
 
 /*
  * routeReducer
@@ -64,27 +62,54 @@ function routeReducer(state = routeInitialState, action) {
   }
 }
 
+const combinedReducer = (history, action) => {
+  if (action.type === "LOGOUT") {
+    return combineReducers({
+      router: connectRouter(history),
+      route: routeReducer,
+      actions: lastActionReducer,
+      login: loginReducer,
+      // Passing an undefined state returns the defaultState
+      navbar: (_state, action) => navbarReducer(undefined, action),
+      overview: (_state, action) => overviewReducer(undefined, action),
+      detailview: (_state, action) => subProjectReducer(undefined, action),
+      workflow: (_state, action) => workflowReducer(undefined, action),
+      workflowitemDetails: (_state, action) => workflowitemDetailsReducer(undefined, action),
+      notifications: (_state, action) => notificationsReducer(undefined, action),
+      documents: (_state, action) => documentsReducer(undefined, action),
+      loading: (_state, action) => loadingReducer(undefined, action),
+      users: (_state, action) => userDashboardReducer(undefined, action),
+      nodes: (_state, action) => nodeDashboardReducer(undefined, action),
+      analytics: (_state, action) => analyticsReducer(undefined, action),
+      confirmation: (_state, action) => confirmationReducer(undefined, action)
+    });
+  } else {
+    return combineReducers({
+      router: connectRouter(history),
+      route: routeReducer,
+      actions: lastActionReducer,
+      navbar: navbarReducer,
+      overview: overviewReducer,
+      detailview: subProjectReducer,
+      workflow: workflowReducer,
+      workflowitemDetails: workflowitemDetailsReducer,
+      notifications: notificationsReducer,
+      login: loginReducer,
+      documents: documentsReducer,
+      loading: loadingReducer,
+      users: userDashboardReducer,
+      nodes: nodeDashboardReducer,
+      analytics: analyticsReducer,
+      confirmation: confirmationReducer
+    });
+  }
+};
+
 /**
  * Creates the main reducer with the asynchronously loaded ones
  */
-export default function createReducer(history) {
-  return combineReducers({
-    router: connectRouter(history),
-    route: routeReducer,
-    actions: lastActionReducer,
-    navbar: navbarReducer,
-    overview: overviewReducer,
-    detailview: subProjectReducer,
-    dashboard: dashboardReducer,
-    workflow: workflowReducer,
-    workflowitemDetails: workflowitemDetailsReducer,
-    notifications: notificationsReducer,
-    login: loginReducer,
-    documents: documentsReducer,
-    loading: loadingReducer,
-    users: userDashboardReducer,
-    nodes: nodeDashboardReducer,
-    analytics: analyticsReducer,
-    confirmation: confirmationReducer
-  });
-}
+const createReducer = history => (state, action) => {
+  return combinedReducer(history, action, state)(state, action);
+};
+
+export default createReducer;